Work on updating the cover mechanism

+++ app/src/main/kotlin/com/kelsos/mbrc/commands/model/UpdateCover.kt
@@ -10,15 +10,20 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.kelsos.mbrc.data.CoverPayload
 import com.kelsos.mbrc.events.bus.RxBus
 import com.kelsos.mbrc.events.ui.CoverChangedEvent
-import com.kelsos.mbrc.extensions.coverFile
+import com.kelsos.mbrc.events.ui.RemoteClientMetaData
+import com.kelsos.mbrc.extensions.getDimens
+import com.kelsos.mbrc.extensions.md5
 import com.kelsos.mbrc.interfaces.ICommand
 import com.kelsos.mbrc.interfaces.IEvent
 import com.kelsos.mbrc.model.MainDataModel
 import com.kelsos.mbrc.services.CoverService
 import com.kelsos.mbrc.widgets.UpdateWidgets
+import com.squareup.picasso.Callback
+import com.squareup.picasso.Picasso
 import rx.Single
 import rx.schedulers.Schedulers
 import timber.log.Timber
+import java.io.File
 import java.io.FileOutputStream
 import javax.inject.Inject
 
@@ -28,22 +33,21 @@ class UpdateCover
                     private val mapper: ObjectMapper,
                     private val coverService: CoverService,
                     private val model: MainDataModel) : ICommand {
+  private val coverDir: File
+
+  init {
+    coverDir = File(context.filesDir, COVER_DIR)
+  }
 
   override fun execute(e: IEvent) {
     val payload = mapper.treeToValue((e.data as JsonNode), CoverPayload::class.java)
 
     if (payload.status == CoverPayload.NOT_FOUND) {
-      clearCover()
+      bus.post(CoverChangedEvent())
       UpdateWidgets.updateCover(context)
     } else if (payload.status == CoverPayload.READY) {
       retrieveCover()
     }
-  }
-
-  private fun clearCover() {
-    val file = context.coverFile()
-    file.delete()
-    bus.post(CoverChangedEvent())
   }
 
   private fun retrieveCover() {
@@ -51,10 +55,13 @@ class UpdateCover
       Single.fromCallable { getBitmap(it) }
     }.flatMap {
       return@flatMap Single.fromCallable { storeCover(it) }
+    }.flatMap {
+      return@flatMap prefetch(it)
     }.subscribeOn(Schedulers.io()).subscribe({
-      bus.post(CoverChangedEvent())
-      model.updateRemoteClient()
-      UpdateWidgets.updateCover(context)
+      model.coverPath = it.absolutePath
+      bus.post(CoverChangedEvent(it.absolutePath))
+      bus.post(RemoteClientMetaData(model.trackInfo, model.coverPath))
+      UpdateWidgets.updateCover(context, it.absolutePath)
     }, {
       removeCover(it)
     })
@@ -71,14 +78,10 @@ class UpdateCover
     } else {
       throw RuntimeException("Base64 was not an image")
     }
-
   }
 
   private fun removeCover(it: Throwable? = null) {
-    val coverFile = context.coverFile()
-    if (coverFile.exists()) {
-      coverFile.delete()
-    }
+    clearPreviousCovers(0)
 
     it?.let {
       Timber.v(it, "Failed to store path")
@@ -87,16 +90,68 @@ class UpdateCover
     bus.post(CoverChangedEvent())
   }
 
-  private fun storeCover(bitmap: Bitmap): String {
-    val file = context.coverFile()
-    file.delete()
+  private fun storeCover(bitmap: Bitmap): File {
+    checkIfExists()
+    clearPreviousCovers()
+    val file = temporaryCover()
     val fileStream = FileOutputStream(file)
     val success = bitmap.compress(JPEG, 100, fileStream)
     if (success) {
-      return file.absolutePath
+      val md5 = file.md5()
+      val extension = file.extension
+      val newFile = File(context.filesDir, "$md5.$extension")
+      file.renameTo(newFile)
+      Timber.v("file was renamed to %s", newFile.absolutePath)
+      return newFile
     } else {
       throw RuntimeException("unable to store cover")
     }
+  }
+
+  private fun prefetch(newFile: File): Single<File> {
+    return Single.create<File> {
+      val dimens = context.getDimens()
+      Picasso.with(context).load(newFile)
+          .config(Bitmap.Config.RGB_565)
+          .resize(dimens, dimens)
+          .centerCrop().fetch(object : Callback {
+        override fun onSuccess() {
+          it.onSuccess(newFile)
+        }
+
+        override fun onError() {
+          it.onError(RuntimeException("Couldn't prefetch cover"))
+        }
+      })
+    }
+  }
+
+  private fun checkIfExists() {
+    if (!coverDir.exists()) {
+      coverDir.mkdir()
+    }
+  }
+
+  private fun clearPreviousCovers(keep: Int = 2) {
+    val storedCovers = coverDir.listFiles()
+    storedCovers.sortByDescending(File::lastModified)
+    val elementsToKeep = if (storedCovers.size - keep < 0) 0 else storedCovers.size - keep
+    storedCovers.takeLast(elementsToKeep).forEach {
+      it.delete()
+    }
+  }
+
+  private fun temporaryCover(): File {
+    val file = File(context.cacheDir, TEMP_COVER)
+    if (file.exists()) {
+      file.delete()
+    }
+    return file
+  }
+
+  companion object {
+    const val COVER_DIR = "cover"
+    const val TEMP_COVER = "temp_cover.jpg"
   }
 
 }
